+++ src/winapp-CLI/WinApp.Cli/Services/IManifestService.cs
@@ -6,6 +6,11 @@ using WinApp.Cli.Models;
 
 namespace WinApp.Cli.Services;
 
+public record AddExecutionAliasOptions(
+    FileInfo ManifestFile,
+    string? AliasName,
+    string? AppId);
+
 public record ManifestGenerationInfo(
     string PackageName,
     string PublisherName,
@@ -37,5 +42,9 @@ internal interface IManifestService
         FileInfo manifestPath,
         FileInfo imagePath,
         TaskContext taskContext,
+        CancellationToken cancellationToken = default);
+
+    public Task<AddExecutionAliasResult> AddExecutionAliasAsync(
+        AddExecutionAliasOptions options,
         CancellationToken cancellationToken = default);
 }
