+++ src/winapp-CLI/WinApp.Cli/Services/ManifestService.cs
@@ -5,7 +5,10 @@ using Spectre.Console;
 using System.Diagnostics;
 using System.Drawing;
 using System.Drawing.Imaging;
+using System.Text;
 using System.Text.RegularExpressions;
+using System.Xml;
+using System.Xml.Linq;
 using WinApp.Cli.ConsoleTasks;
 using WinApp.Cli.Helpers;
 using WinApp.Cli.Models;
@@ -465,4 +468,244 @@ internal partial class ManifestService(
 
     [GeneratedRegex(@"(\d+)x(\d+)", RegexOptions.IgnoreCase)]
     private static partial Regex DimensionRegex();
+
+    private static readonly XNamespace AppxDefaultNs = "http://schemas.microsoft.com/appx/manifest/foundation/windows10";
+    private static readonly XNamespace Uap5Ns = "http://schemas.microsoft.com/appx/manifest/uap/windows10/5";
+
+    [GeneratedRegex(@"^(\s*)<([\w:.-]+)((?:\s+[\w:.-]+\s*=\s*""[^""]*"")+)\s*(\/?>)\s*$")]
+    private static partial Regex TagPattern();
+
+    [GeneratedRegex(@"([\w:.-]+\s*=\s*""[^""]*"")")]
+    private static partial Regex AttrPattern();
+
+    public async Task<AddExecutionAliasResult> AddExecutionAliasAsync(
+        AddExecutionAliasOptions options,
+        CancellationToken cancellationToken = default)
+    {
+        XDocument doc;
+        try
+        {
+            doc = XDocument.Load(options.ManifestFile.FullName);
+        }
+        catch (Exception ex)
+        {
+            return new AddExecutionAliasResult(AddExecutionAliasStatus.ManifestParseError, ErrorMessage: ex.Message);
+        }
+
+        var root = doc.Root;
+        if (root == null)
+        {
+            return new AddExecutionAliasResult(AddExecutionAliasStatus.ManifestEmpty);
+        }
+
+        // Find the target Application element
+        var applications = root.Descendants(AppxDefaultNs + "Application").ToList();
+        if (applications.Count == 0)
+        {
+            return new AddExecutionAliasResult(AddExecutionAliasStatus.NoApplicationElement);
+        }
+
+        XElement targetApp;
+        if (!string.IsNullOrEmpty(options.AppId))
+        {
+            targetApp = applications.FirstOrDefault(a =>
+                string.Equals(a.Attribute("Id")?.Value, options.AppId, StringComparison.OrdinalIgnoreCase))!;
+            if (targetApp == null)
+            {
+                return new AddExecutionAliasResult(AddExecutionAliasStatus.ApplicationIdNotFound);
+            }
+        }
+        else
+        {
+            targetApp = applications[0];
+        }
+
+        // Infer alias name from Executable attribute if not specified
+        var aliasName = options.AliasName;
+        if (string.IsNullOrEmpty(aliasName))
+        {
+            var executable = targetApp.Attribute("Executable")?.Value;
+            if (!string.IsNullOrEmpty(executable))
+            {
+                aliasName = executable;
+            }
+            else
+            {
+                return new AddExecutionAliasResult(AddExecutionAliasStatus.CouldNotInferAlias);
+            }
+        }
+
+        // Ensure alias ends with .exe
+        if (!aliasName.EndsWith(".exe", StringComparison.OrdinalIgnoreCase))
+        {
+            aliasName += ".exe";
+        }
+
+        // Check if the target Application already has any execution alias
+        var targetExtensions = targetApp.Element(AppxDefaultNs + "Extensions");
+        if (targetExtensions != null)
+        {
+            var existingAliasElements = targetExtensions
+                .Elements(Uap5Ns + "Extension")
+                .Where(e => string.Equals(e.Attribute("Category")?.Value, "windows.appExecutionAlias", StringComparison.OrdinalIgnoreCase))
+                .Descendants(Uap5Ns + "ExecutionAlias")
+                .Select(e => e.Attribute("Alias")?.Value)
+                .Where(v => v != null)
+                .ToList();
+
+            if (existingAliasElements.Count > 0)
+            {
+                var existingAlias = existingAliasElements[0]!;
+                if (string.Equals(existingAlias, aliasName, StringComparison.OrdinalIgnoreCase))
+                {
+                    return new AddExecutionAliasResult(AddExecutionAliasStatus.AlreadyExists, AliasName: aliasName);
+                }
+                else
+                {
+                    return new AddExecutionAliasResult(AddExecutionAliasStatus.ConflictingAliasExists, AliasName: aliasName, ExistingAlias: existingAlias);
+                }
+            }
+        }
+
+        // Ensure uap5 namespace is declared on the Package element
+        if (root.GetNamespaceOfPrefix("uap5") == null)
+        {
+            root.Add(new XAttribute(XNamespace.Xmlns + "uap5", Uap5Ns));
+        }
+
+        // Ensure uap5 is in IgnorableNamespaces
+        var ignorableAttr = root.Attribute("IgnorableNamespaces");
+        if (ignorableAttr != null)
+        {
+            var namespaces = ignorableAttr.Value.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+            if (!namespaces.Contains("uap5", StringComparer.OrdinalIgnoreCase))
+            {
+                ignorableAttr.Value = ignorableAttr.Value + " uap5";
+            }
+        }
+
+        // Build the ExecutionAlias element
+        var aliasElement = new XElement(Uap5Ns + "ExecutionAlias",
+            new XAttribute("Alias", aliasName));
+
+        // Find or create the Extensions > uap5:Extension > uap5:AppExecutionAlias hierarchy
+        var extensions = targetApp.Element(AppxDefaultNs + "Extensions");
+        if (extensions == null)
+        {
+            extensions = new XElement(AppxDefaultNs + "Extensions");
+            targetApp.Add(extensions);
+        }
+
+        // Look for an existing uap5:Extension with Category="windows.appExecutionAlias"
+        var aliasExtension = extensions.Elements(Uap5Ns + "Extension")
+            .FirstOrDefault(e => string.Equals(
+                e.Attribute("Category")?.Value,
+                "windows.appExecutionAlias",
+                StringComparison.OrdinalIgnoreCase));
+
+        if (aliasExtension != null)
+        {
+            // Add to existing AppExecutionAlias block
+            var appExecAlias = aliasExtension.Element(Uap5Ns + "AppExecutionAlias");
+            if (appExecAlias != null)
+            {
+                appExecAlias.Add(aliasElement);
+            }
+            else
+            {
+                var newAppExecAlias = new XElement(Uap5Ns + "AppExecutionAlias", aliasElement);
+                aliasExtension.Add(newAppExecAlias);
+            }
+        }
+        else
+        {
+            // Create new Extension block
+            var newExtension = new XElement(Uap5Ns + "Extension",
+                new XAttribute("Category", "windows.appExecutionAlias"),
+                new XElement(Uap5Ns + "AppExecutionAlias", aliasElement));
+            extensions.Add(newExtension);
+        }
+
+        // Save with UTF-8 no BOM and proper indentation
+        var utf8NoBom = new UTF8Encoding(encoderShouldEmitUTF8Identifier: false);
+        var settings = new XmlWriterSettings
+        {
+            Indent = true,
+            IndentChars = "  ",
+            Encoding = utf8NoBom,
+            OmitXmlDeclaration = doc.Declaration == null,
+        };
+
+        // Write to memory first so we can post-process attribute formatting
+        string xmlContent;
+        using (var memoryStream = new MemoryStream())
+        {
+            using (var writer = XmlWriter.Create(memoryStream, settings))
+            {
+                doc.Save(writer);
+            }
+
+            xmlContent = utf8NoBom.GetString(memoryStream.ToArray());
+        }
+
+        // Split attributes onto separate lines for elements with more than 2 attributes
+        xmlContent = FormatXmlAttributes(xmlContent);
+
+        await File.WriteAllTextAsync(options.ManifestFile.FullName, xmlContent, utf8NoBom, cancellationToken);
+
+        return new AddExecutionAliasResult(AddExecutionAliasStatus.Added, AliasName: aliasName);
+    }
+
+    /// <summary>
+    /// Post-processes XML output to place each attribute on its own line
+    /// when an element has more than 2 attributes, improving readability.
+    /// </summary>
+    internal static string FormatXmlAttributes(string xml)
+    {
+        var result = new StringBuilder();
+
+        foreach (var rawLine in xml.Split('\n'))
+        {
+            var line = rawLine.TrimEnd('\r');
+            var match = TagPattern().Match(line);
+            if (match.Success)
+            {
+                var indent = match.Groups[1].Value;
+                var tagName = match.Groups[2].Value;
+                var attrsStr = match.Groups[3].Value;
+                var closing = match.Groups[4].Value;
+
+                var attrs = AttrPattern().Matches(attrsStr);
+                if (attrs.Count > 2)
+                {
+                    var attrIndent = indent + "  ";
+                    result.Append(indent).Append('<').Append(tagName);
+                    foreach (Match attr in attrs)
+                    {
+                        result.Append(Environment.NewLine).Append(attrIndent).Append(attr.Value.Trim());
+                    }
+
+                    result.Append(closing == "/>" ? " />" : ">");
+                    result.Append(Environment.NewLine);
+                }
+                else
+                {
+                    result.Append(line).Append(Environment.NewLine);
+                }
+            }
+            else
+            {
+                result.Append(line).Append(Environment.NewLine);
+            }
+        }
+
+        // Trim the trailing extra newline added by the loop
+        var newLine = Environment.NewLine;
+        if (result.Length >= newLine.Length)
+        {
+            result.Length -= newLine.Length;
+        }
+
+        return result.ToString();
+    }
 }
